backend/requirements.txt
flask==3.1.2
flask-cors==6.0.2
flask-sqlalchemy==3.1.1
requests==2.32.5
psycopg2-binary==2.9.11
flask-socketio==5.3.6
python-socketio==5.11.0
polyline==1.4.0

# Authentication dependencies
flask-jwt-extended==4.6.0
bcrypt==4.2.1
flask-mail==0.9.1
email-validator==2.2.0
python-dotenv==1.0.1
itsdangerous==2.2.0
